GlobalPerformanceLogger: handle cases when global.performance is not yet initialized (#37918)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/37918

# Changelog:
[Internal] -

There is no guarantee that `setUpPeformance`, which establishes the `global.performance` object accordingly, will be always called before anything tries to use performance logging.

In fact, there is an example in FB4A with the early initialization of `MobileConfige`, which logs a time span to measure how long does it take to initialize.

This makes the web performance logging code safe from such scenarios.

Reviewed By: NickGerleman

Differential Revision: D46762527

fbshipit-source-id: e189cc3f587ade31a2b5d853634c35da207f5faa