Add plumbing/boilerplate for an iOS implementation of the `gotpointercapture` and `lostpointercapture` events (second try) (#37889)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/37889

Changelog: [Internal] - Add plumbing/boilerplate for an iOS implementation of the gotpointercapture and lostpointercapture events

Lets try this again: The key difference between this and D44977499 (which I previously reverted) is that in propsConversions & primitives I've ommited the "capture" versions of those methods as it was causing issues. Since we're not doing any runtime checks of those raw props it isn't particularlly necessary (at least not yet) and if we ever want to we can address that when it comes up.

The original diff description follows:

This diff simply adds the boilerplate necessary to hook up the gotpointercapture and lostpointercapture events to the fabric iOS touch handler. This diff does not contain any actual implementation of their behavior as that will occur in future diffs.

Reviewed By: adanoff

Differential Revision: D46709127

fbshipit-source-id: 339dc99e14f2e72d9116bbe84edc67cb5bfaa73b