Add partner application process to ECOSYSTEM.md (#37689)
Summary:
## Overview

This PR updates the ECOSYSTEM.md with more information for how to apply to be a partner, and includes an update to the joining process including a 3 month incubation period.

It also emphasizes that being a partner is a commitment of resources to improve React Native, and not a status symbol.

Changelog: internal

Pull Request resolved: https://github.com/facebook/react-native/pull/37689

Reviewed By: lunaleaps, cortinico

Differential Revision: D46941950

Pulled By: rickhanlonii

fbshipit-source-id: 254b0927d5e57a5169c52ea7ac5eaff1d064ab2b