Guard against stopped surface in FabricUIManager.getColor (#38239)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/38239

In D46840717 I made SurfaceMountingManager reset its context on stop as the Context may be keeping various other objects alive and was causing memory leaks.

This does cause a race condition if JS is still rendering at this time, as it needs to access to context to support things like PlatformColor.

Changelog: [Android] Fix crash on teardown in new architecture

Reviewed By: rshest

Differential Revision: D47290581

fbshipit-source-id: ef83c1d110d1156e39f2491fb62c82ec41c3b8e1